rename: view 패키지 통합
